Correct problem with incorrectly capped names
A problem has existed in the /perms command that causes a local players UUID (in offline mode) to be changed, causing them to lose inventory, etc.  The problem comes when someone runs a command like `/perms users suresttexas00 info`.  lookupUsername() would take the improperly capitalized name and pass it to formatUUID, changing the player's local name from "SurestTexas00'" to "suresttexas00", causing the offline server to think it is a new player. This code reads the properly capped name from Mojang and returns that in lieu of the inadvertently improperly capped supplied name.