update Packet read and send methods read and send are now wrapped abstractions for readpkt and sendpkt.
readpkt and sendpkt are lower-level methods that eliminate the use of
string arguments like "double:xposition|double:yposition|double:zposition|bool:on_ground".

Arguments for read (like the example cited) are reduced simply to number codes (constants)
for the desired data types to be returned (i.e., 8 for "double").  Instead, in the example
cited, the readpkt argument would simply be '[8, 8, 8, 10]'.  Of course, the numbers can be
abstracted with constants for readability.  read() returns a list of the results in the
same order as the numbers/constants were passed.  The reading function can decide what variable names
to assign based on their list position:

```
data = packet.newread(_DOUBLE, _DOUBLE, _DOUBLE, _BOOL)
x, y, z, on_ground = data
```

similarly, sendpkt() is a bit like the original send(), except that the
field types are abstracted numbers too and don't need the slicing operations to get the
arguments (the payload portion does not change).

The primary aim of these changes is to remove un-needed string operations while preserving
readability in the code.  This code is run in high frequency on a packet stream.. The
unneeded string parsing operations and the overly fancy use of slower dictionary methods
needs to be limited in this high-use code to improve wrapper's network speed.