Finish all send() and read() references to [send/read]pkt() - added an interesting snippet into helpers for possible later usage with player.dat files - all references to read() and send() are corrected. - beefed up player inventory to be refreshed when other windows are open besides the inventory.
There is one pernicious problem with wrapper inventories still... It occurs when the client
moves an item (this is easy to see when you use a gold shovel or wood ax or other commonly used
 "wand" item) from one inventory slot to another. wrapper has no parsing for a client-side only
movement of an item from, say, #36 position to #40.  Even though the clientbound set_slot and
open_window items on serverconnection are beefed up a little and do other windows besides
the inventory (0)... they only add the new location and dont delete the old one.. resulting in
"two" wands, for instance in wrapper's inventory... so we are not much better off.

Need a way to delete the old one that was moved and find out how client is communicating
the slot info to the server (server has the info some how), and start parsing it. maybe from
pktSB.CLICK_WINDOW?